[Refactor] Modify package name of utils to keep consistence
