+++ client/src/main/java/com/aliyun/emr/rss/client/read/DfsPartitionReader.java
@@ -41,7 +41,7 @@ import com.aliyun.emr.rss.common.network.protocol.Message;
 import com.aliyun.emr.rss.common.network.protocol.OpenStream;
 import com.aliyun.emr.rss.common.protocol.PartitionLocation;
 import com.aliyun.emr.rss.common.util.Utils;
-import com.aliyun.emr.rss.common.utils.ShuffleBlockInfoUtils;
+import com.aliyun.emr.rss.common.util.ShuffleBlockInfoUtils;
 
 public class DfsPartitionReader implements PartitionReader {
   private final int chunkSize;
