+++ common/src/main/java/com/aliyun/emr/rss/common/util/ShuffleBlockInfoUtils.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.aliyun.emr.rss.common.utils;
+package com.aliyun.emr.rss.common.util;
 
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
