+++ server-worker/src/main/java/com/aliyun/emr/rss/service/deploy/worker/LevelDBProvider.java
@@ -29,7 +29,7 @@ import org.iq80.leveldb.Options;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.aliyun.emr.rss.common.utils.PBSerDeUtils;
+import com.aliyun.emr.rss.common.util.PBSerDeUtils;
 
 /**
  * LevelDB utility class available in the network package.
