+++ server-worker/src/main/java/com/aliyun/emr/rss/service/deploy/worker/storage/PartitionFilesSorter.java
@@ -54,9 +54,9 @@ import com.aliyun.emr.rss.common.network.server.MemoryTracker;
 import com.aliyun.emr.rss.common.unsafe.Platform;
 import com.aliyun.emr.rss.common.util.ThreadUtils;
 import com.aliyun.emr.rss.common.util.Utils;
-import com.aliyun.emr.rss.common.utils.PBSerDeUtils;
-import com.aliyun.emr.rss.common.utils.ShuffleBlockInfoUtils;
-import com.aliyun.emr.rss.common.utils.ShuffleBlockInfoUtils.ShuffleBlockInfo;
+import com.aliyun.emr.rss.common.util.PBSerDeUtils;
+import com.aliyun.emr.rss.common.util.ShuffleBlockInfoUtils;
+import com.aliyun.emr.rss.common.util.ShuffleBlockInfoUtils.ShuffleBlockInfo;
 import com.aliyun.emr.rss.service.deploy.worker.LevelDBProvider;
 import com.aliyun.emr.rss.service.deploy.worker.ShuffleRecoverHelper;
 import com.aliyun.emr.rss.service.deploy.worker.WorkerSource;
